refactor(settings): restructure and split multi-account into components
