+++ __tests__/screens/settings-screen/settings.spec.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { render, fireEvent, waitFor, screen } from "@testing-library/react-native"
 import { loadLocale } from "@app/i18n/i18n-util.sync"
 import { i18nObject } from "@app/i18n/i18n-util"
-import { SwitchAccountComponent } from "@app/screens/settings-screen/settings/swiitch-account.stories"
+import { SwitchAccountComponent } from "@app/screens/settings-screen/settings/multi-account/switch-account.stories"
 import { GetUsernamesDocument, SettingsScreenDocument } from "@app/graphql/generated"
 import KeyStoreWrapper from "@app/utils/storage/secureStorage"
 import mocks from "@app/graphql/mocks"
