+++ app/screens/settings-screen/settings/multi-account/index.ts
@@ -1,0 +1,1 @@
+export * from "./switch-account"
