+++ app/screens/settings-screen/settings/multi-account/switch-account.stories.tsx
@@ -1,8 +1,8 @@
 import * as React from "react"
-
 import { MockedProvider, MockedResponse } from "@apollo/client/testing"
-import { createCache } from "../../../graphql/cache"
-import { IsAuthedContextProvider } from "../../../graphql/is-authed-context"
+
+import { createCache } from "../../../../graphql/cache"
+import { IsAuthedContextProvider } from "../../../../graphql/is-authed-context"
 import { SwitchAccount } from "./switch-account"
 
 export const SwitchAccountComponent = ({ mock }: { mock: MockedResponse[] }) => (
